requirements.txt
agora==1.2.3
argh==0.26.2
asgiref==3.5.0
astroid==2.9.3
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==35.0.0
dateutils==0.6.12
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-apptemplates==1.5
django-axes==5.31.0
django-bootstrap5==21.3
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-ipware==4.0.2
django-stdimage==5.3.0
django-summernote==0.8.20.0
filelock==3.4.0
Grafico==0.1.3
gunicorn==20.1.0
h11==0.12.0
heroku==0.1.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.22.3
onyx.core==1.0.4
outcome==1.1.0
packages==0.1.0
packaging==21.3
packed==0.2.0
pbr==5.9.0
Pillow==9.1.0
pipenv==2021.11.23
platformdirs==2.4.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
requests==2.27.1
scipy==1.8.0
scripts==2.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tzdata==2022.1
virtualenv==20.10.0
